requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
base58==2.1.1
bip-utils==2.7.0
bitmerchant==0.1.8
cachetools==5.3.0
cbor2==5.4.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coincurve==17.0.0
crcmod==1.7
Django==4.2.1
djangorestframework==3.14.0
ecdsa==0.18.0
ed25519-blake2b==1.4
gunicorn==20.1.0
hdwallet==0.1.0a1
idna==3.4
install==1.3.5
mnemonic==0.20
mock==5.0.2
model-bakery==1.11.0
parameterized==0.9.0
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.17
PyNaCl==1.5.0
pysha3==1.0.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
